move LoginPage from container to component
